// eid-pdeau/src/eID.PDEAU.UI/src/app/features/centers/centers.dto.ts
export const PageableSchema = z.object({
    sort: z.any(),
    offset: z.number(),
    pageNumber: z.number(),
    pageSize: z.number(),
    paged: z.boolean(),
    unpaged: z.boolean(),
});

export const CenterApplicationBasicSchema = z.object({
    applicationNumber: z.string(),
    applicationType: z.nativeEnum(ApplicationType),
    eidName: z.string(),
    status: z.nativeEnum(ApplicationStatus),
    id: z.string(),
    createDate: z.string(),
});

export const CentersPageableSchema = z.object({
    content: z.array(CenterApplicationBasicSchema),
    pageable: PageableSchema,
    last: z.boolean(),
    totalElements: z.number(),
    totalPages: z.number(),
    size: z.number(),
    number: z.number(),
    sort: z.any(),
    first: z.boolean(),
    numberOfElements: z.number(),
    empty: z.boolean(),
});

export const documentTypeSchema = z.object({
    id: z.string().uuid(),
    name: z.string(),
    active: z.boolean(),
    requiredForAdministrator: z.boolean().nullish(),
    requiredForCenter: z.boolean().nullish(),
    descriptions: z.record(z.string(), z.any()),
});

export const documentTypeListSchema = z.array(documentTypeSchema);

export type documentType = z.infer<typeof documentTypeSchema>;

export const EmployeeSchema = z.object({
    roles: z.array(z.string()),
    name: z.string(),
    nameLatin: z.string(),
    phoneNumber: z.string(),
    email: z.string(),
    citizenIdentifierType: z.string(),
    citizenIdentifierNumber: z.string(),
});

export const EidManagerFrontOfficeSchema = z.object({
    name: z.string(),
    location: z.string(),
    region: z.string(),
    contact: z.string(),
    longitude: z.string(),
    latitude: z.string(),
    workingHours: z.array(z.string()),
    email: z.string(),
    description: z.string(),
});

export const PersonSchema = z.object({
    name: z.string(),
    nameLatin: z.string(),
    phoneNumber: z.string().nullish(),
    email: z.string(),
    citizenIdentifierType: z.string(),
    citizenIdentifierNumber: z.string(),
});

export const AttachmentSchema = z.object({
    fileName: z.string(),
    content: z.string(),
    documentType: z.string(),
});

export type Attachment = z.infer<typeof AttachmentSchema>;

export const AttachmentFullSchema = z.object({
    id: z.string().uuid(),
    fileName: z.string(),
    content: z.string(),
    documentType: documentTypeSchema,
    isOutgoing: z.boolean().nullish(),
    createDate: z.string(),
});

export type AttachmentFull = z.infer<typeof AttachmentFullSchema>;

export const AttachmentsWithNotesSchema = z.object({
    attachments: z.array(AttachmentFullSchema),
    notes: z.array(z.any()).nullish(),
});

export const ApplicationSchema = z.object({
    employees: z.array(EmployeeSchema),
    eidManagerFrontOffices: z.array(EidManagerFrontOfficeSchema),
    authorizedPersons: z.array(PersonSchema),
    applicant: PersonSchema,
    applicationType: z.string(),
    attachments: z.array(AttachmentSchema),
    homePage: z.string(),
    companyName: z.string(),
    companyNameLatin: z.string(),
    eikNumber: z.string(),
    address: z.string(),
    email: z.string(),
    phone: z.string(),
    description: z.string(),
    logoUrl: z.string().optional(),
    applicationNumber: z.object({
        id: z.string(),
        applicationType: z.string(),
    }).optional(),
});

export type CenterApplication = z.infer<typeof ApplicationSchema>;

export const ApplicationDetailsSchema = z.object({
    id: z.string(),
    referenceNumber: z.string().nullish(),
    employees: z.array(EmployeeSchema),
    eidManagerFrontOffices: z.array(EidManagerFrontOfficeSchema),
    authorizedPersons: z.array(PersonSchema),
    applicant: PersonSchema,
    applicationType: z.string(),
    attachments: z.array(AttachmentFullSchema),
    homePage: z.string().nullish(),
    companyName: z.string(),
    companyNameLatin: z.string(),
    eikNumber: z.string(),
    address: z.string(),
    email: z.string(),
    phone: z.string().nullish(),
    description: z.string().nullish(),
    status: z.nativeEnum(ApplicationStatus),
    notes: z.array(z.any()),
    applicationNumber: z.object({
        id: z.string(),
        applicationType: z.string(),
    })
});
// ...
export const RegionSchema = z.object({
    code: z.string(),
    descriptions: z.object({
        en: z.string(),
        bg: z.string(),
    }),
});

export type Region = z.infer<typeof RegionSchema>;

export const RegionListSchema = z.array(RegionSchema);

const EidFrontOfficeSchema = z.object({
    id: z.string().uuid(),
    name: z.string(),
    region: z.string(),
});

export const CenterSchema = z.object({
    id: z.string().uuid(),
    name: z.string(),
    nameLatin: z.string().optional(),
    eikNumber: z.string(),
    homePage: z.string(),
    logoUrl: z.string(),
    address: z.string(),
    email: z.string().email(),
    code: z.string(),
    serviceType: z.any(),
    managerStatus: z.nativeEnum(ManagerStatus),
    authorizedPersons: z.array(PersonSchema),
    employeeIds: z.array(z.string()),
    eidManagerFrontOfficeIds: z.array(z.any()),
    attachmentIds: z.array(z.string()),
    noteIds: z.union([z.null(), z.array(z.any())]),
    providedServiceIds: z.array(z.any()),
});